+++ include/zenoh-pico/transport/link/tx.h
@@ -21,19 +21,19 @@
 
 void __unsafe_z_prepare_wbuf(_z_wbuf_t *buf, int is_streamed);
 void __unsafe_z_finalize_wbuf(_z_wbuf_t *buf, int is_streamed);
-int __unsafe_z_serialize_zenoh_fragment(_z_wbuf_t *dst, _z_wbuf_t *src, z_reliability_t reliability, size_t sn);
+int8_t __unsafe_z_serialize_zenoh_fragment(_z_wbuf_t *dst, _z_wbuf_t *src, z_reliability_t reliability, size_t sn);
 _z_transport_message_t _z_frame_header(z_reliability_t reliability, int is_fragment, int is_final, _z_zint_t sn);
 
 /*------------------ Transmission and Reception helpers ------------------*/
-int _z_unicast_send_z_msg(_z_session_t *zn, _z_zenoh_message_t *z_msg, z_reliability_t reliability,
-                          z_congestion_control_t cong_ctrl);
-int _z_multicast_send_z_msg(_z_session_t *zn, _z_zenoh_message_t *z_msg, z_reliability_t reliability,
-                            z_congestion_control_t cong_ctrl);
+int8_t _z_unicast_send_z_msg(_z_session_t *zn, _z_zenoh_message_t *z_msg, z_reliability_t reliability,
+                             z_congestion_control_t cong_ctrl);
+int8_t _z_multicast_send_z_msg(_z_session_t *zn, _z_zenoh_message_t *z_msg, z_reliability_t reliability,
+                               z_congestion_control_t cong_ctrl);
 
-int _z_send_t_msg(_z_transport_t *zt, const _z_transport_message_t *t_msg);
-int _z_unicast_send_t_msg(_z_transport_unicast_t *ztu, const _z_transport_message_t *t_msg);
-int _z_multicast_send_t_msg(_z_transport_multicast_t *ztm, const _z_transport_message_t *t_msg);
+int8_t _z_send_t_msg(_z_transport_t *zt, const _z_transport_message_t *t_msg);
+int8_t _z_unicast_send_t_msg(_z_transport_unicast_t *ztu, const _z_transport_message_t *t_msg);
+int8_t _z_multicast_send_t_msg(_z_transport_multicast_t *ztm, const _z_transport_message_t *t_msg);
 
-int _z_link_send_t_msg(const _z_link_t *zl, const _z_transport_message_t *t_msg);
+int8_t _z_link_send_t_msg(const _z_link_t *zl, const _z_transport_message_t *t_msg);
 
 #endif /* ZENOH_PICO_TRANSPORT_LINK_TX_H */
